Move Negotiation SSM config to a separate package
